Fix compilation on older gcc's, roll ProjectorCacheKey into projector.cc
